backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
temporalio==1.5.0
google-generativeai==0.3.2
pypdf==3.17.0
google-adk==1.21
